課題11（後半）

+++ app/admin/categories/[id]/page.tsx
@@ -2,6 +2,7 @@
 import { useState, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import CategoryForm from "@/app/_components/Form/CategoryForm";
+import { useSupabaseSession } from "@/_hooks/useSupabaseSession";
 
 type Category = {
   id: number;
@@ -17,7 +18,7 @@ type Post = {
   title: string;
   createdAt: string;
   content: string;
-  thumbnailUrl: string;
+  thumbnailImageKey: string;
   postCategories: PostCategory[]
 }
 
@@ -30,21 +31,29 @@ type Props = {
 
 
 export default function AdminEditCategoryPage({ params }: Props) {
+  const { token } = useSupabaseSession();
   const categoryId = params.id
   const router = useRouter();
 
   const [name, setName] = useState<string>("");
 
   useEffect(() => {
+    if (!token) return
     const fetchName = async () => {
-      const res = await fetch(`/api/admin/categories/${categoryId}`)
+      const res = await fetch(`/api/admin/categories/${categoryId}`, {
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: token,
+        }
+      })
       const data = await res.json();
       setName(data.category.name);
     }
     fetchName()
-  }, [categoryId])
+  }, [categoryId, token])
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    if(!token)return
     e.preventDefault();
     const formData = new FormData(e.currentTarget);
 
@@ -54,7 +63,10 @@ export default function AdminEditCategoryPage({ params }: Props) {
 
     await fetch(`/api/admin/categories/${categoryId}`, {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: token,
+      },
       body: JSON.stringify(payload),
     })
     alert("カテゴリーを更新しました")
@@ -62,10 +74,15 @@ export default function AdminEditCategoryPage({ params }: Props) {
   }
 
   const handleDelete = async () => {
+    if(!token)return
     const ok = window.confirm("本当に削除しますか？")
     if (!ok) return;
     await fetch(`/api/admin/categories/${categoryId}`, {
       method: "DELETE",
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: token,
+      },
     });
     alert("カテゴリーを削除しました");
     router.push("/admin/categories")
@@ -75,7 +92,7 @@ export default function AdminEditCategoryPage({ params }: Props) {
 
   return (
     <>
-      <CategoryForm title="カテゴリー編集" defaultValue={name} onSubmit={handleSubmit} onDelete={handleDelete} submitLabel="更新"/>
+      <CategoryForm title="カテゴリー編集" defaultValue={name} onSubmit={handleSubmit} onDelete={handleDelete} submitLabel="更新" />
     </>
   )
