+++ app/admin/posts/[id]/page.tsx
@@ -20,6 +20,7 @@
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import PostForm from "@/app/_components/Form/PostForm";
+import { useSupabaseSession } from "@/_hooks/useSupabaseSession";
 
 /* ===== 型定義 ===== */
 
@@ -37,7 +38,7 @@ type Post = {
   title: string;
   createdAt: string;
   content: string;
-  thumbnailUrl: string;
+  thumbnailImageKey: string;
   postCategories: PostCategory[]
 }
 
@@ -62,13 +63,14 @@ export default function AdminEditPostPage({ params }: Props) {
   const router = useRouter();
   //JavaScriptからページ遷移をするための道具を取得している。後に記載しているrouter.push("/admin/posts");で更新後にページ遷移したいため
   const postId = params.id;
+  const { token } = useSupabaseSession();
 
 
 
   /* ===== 初期表示用データ ===== */
   const [title, setTitle] = useState<string>("");
   const [content, setContent] = useState<string>("");
-  const [thumbnailUrl, setThumbnailUrl] = useState<string>("");
+  const [thumbnailImageKey, setThumbnailImageKey] = useState<string>("");
 
 
   /* ===== カテゴリー選択状態 ===== */
@@ -77,20 +79,26 @@ export default function AdminEditPostPage({ params }: Props) {
 
   /* ===== 記事詳細取得 ===== */
   useEffect(() => {
+    if (!token) return
     const fetchPost = async () => {
-      const res = await fetch(`/api/admin/posts/${postId}`)
+      const res = await fetch(`/api/admin/posts/${postId}`, {
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: token,
+        }
+      })
       const json = await res.json();
       const data: Post = json.post;
       setTitle(data.title);
       setContent(data.content);
-      setThumbnailUrl(data.thumbnailUrl);
+      setThumbnailImageKey(data.thumbnailImageKey);
       setSelectedCategories(
         data.postCategories.map((pc) => pc.category.id)
       )
       //PostCategory[] → number[]にデータ整形している作業
     };
     fetchPost();
-  }, [postId])
+  }, [postId, token])
 
   /* ===== チェックボックス切替 ===== */
   const toggleCategory = (id: number) => {
@@ -104,19 +112,23 @@ export default function AdminEditPostPage({ params }: Props) {
   /* ===== 更新処理 ===== */
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
+    if (!token) return
 
     const formData = new FormData(e.currentTarget);
 
     const payload = {
       title: formData.get("title") as string,
       content: formData.get("content") as string,
-      thumbnailUrl: formData.get("thumbnailUrl") as string,
+      thumbnailImageKey,
       categories: selectedCategories.map((id) => ({ id })),
     };
 
     await fetch(`/api/admin/posts/${postId}`, {
       method: "PUT",
-      headers: { "Content-Type": "application/json" },
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: token,
+      },
       body: JSON.stringify(payload),
     });
 
@@ -125,10 +137,15 @@ export default function AdminEditPostPage({ params }: Props) {
   };
 
   const handleDelete = async () => {
+    if(!token)return
     const ok = window.confirm("本当に削除しますか？");
     if (!ok) return;
     await fetch(`/api/admin/posts/${postId}`, {
       method: "DELETE",
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: token,
+      },
     });
     alert("記事を削除しました");
     router.push("/admin/posts");
@@ -143,11 +160,12 @@ export default function AdminEditPostPage({ params }: Props) {
       initialValues={{
         title,
         content,
-        thumbnailUrl,
+        thumbnailImageKey,
       }}
       selectedCategories={selectedCategories}
       onToggleCategory={toggleCategory}
       submitLabel="更新"
+      onImageUploaded={setThumbnailImageKey} 
     />
   )
